refactor(derive): split context.rs into context/ module directory
Split the 1884-line context.rs (71.5KB) into three focused submodules:
- context/helpers.rs (~55 lines): cmd_base_key(), format_task_line()
- context/session.rs (~500 lines code + tests): session history rendering
- context/mod.rs (~200 lines code + tests): render_context() entry point

Public API unchanged — lib.rs still exports render_context via mod context.
All 43 tests preserved and passing, zero clippy warnings.

Closes #294
